Add files via upload
IoT Hacking – This project was about IoT Hacking of a Smart House System in various situation, which took control by me (within the scene for the coursework purpose only), and found all the loopholes, which generally hackers and hacktivist group uses.

For this course, the course superior was Dr. Basil Elmasri.

For this Hacking, I have used Linux/ Unix OS.

For more detail, and to understand each stage, look inside the hacking activities.